Update thêm button edit vào bảng

+++ www/app/app.js
@@ -50,24 +50,36 @@ angular.module('starter', ['ionic', 'ngCordova'])
         }
         else {
             var myPopup = $ionicPopup.show({
-                title: 'Nhập thông tin sai!!!'
+                title: 'Nhập thiếu thông tin, vui lòng nhập đầy đủ!!!'
             });
             $timeout(function() {
                 myPopup.close();
-            }, 500);
+            }, 1000);
         }
     }
     // Xóa dữ liệu trong bảng
     $scope.del = function(product) {
-        var query = "DELETE FROM fruit";
-        $cordovaSQLite.execute(db, query);
-        var myPopup = $ionicPopup.show({
-            title: 'Xóa thành công!!!'
+        var confirmPopup = $ionicPopup.confirm({
+            title: 'Cảnh báo',
+            template: 'Bạn có muốn xóa dữ liệu?'
         });
-        $timeout(function() {
-            myPopup.close();
-        }, 500);
-        $scope.show();
+        confirmPopup.then(function(res) {
+            if(res) {
+                var query = "DELETE FROM fruit";
+                $cordovaSQLite.execute(db, query);
+                var myPopup = $ionicPopup.show({
+                    title: 'Xóa thành công!!!'
+                });
+                $timeout(function() {
+                    myPopup.close();
+                }, 500);
+                $scope.show();
+            }
+        });
+    }
+    // Chỉnh sửa
+    $scope.edit = function(product) {
+        $scope.showPopup();
     }
     // Hiển thị dữ liệu
     $scope.show = function() {
@@ -83,4 +95,31 @@ angular.module('starter', ['ionic', 'ngCordova'])
             console.log(err);
         });
     }
+
+    $scope.showPopup = function() {
+        var myPopup = $ionicPopup.show({
+            template: '<input type="text" ng-model="product.name">',
+            title: 'Chỉnh sửa thông tin',
+            scope: $scope,
+            inputPlaceholder: 'Your password',
+            buttons: [
+                { text: 'Cancel' },
+                {
+                    text: '<b>Save</b>',
+                    type: 'button-positive',
+                    onTap: function(e) {
+                        if (!$scope.product.name) {
+                            e.preventDefault();
+                        } else {
+                            return $scope.product.name;
+                        }
+                    }
+                },
+                { text: 'Help' },
+            ]
+        });
+        myPopup.then(function(res) {
+            console.log('Tapped!', res);
+        });
+    };
 })
